Initial Commit
Here we've got an initial version of the xmem driver,
almost fully CDCM-compliant.

Known issues at the moment:
- CDCM has to be fixed to support caching the ISRs and therefore
to avoid ugly ifdefs when defining the interrupt handler (this has
to be done in the driver in LynxOs' fashion and everything has to
be transparent to the driver writer)--this is why above I use the
words 'almost fully CDCM-compliant'.

- There is an improvement to be made to adapt to the model of the
timing system we're aiming for; an interrupt mask should be added
for the clients so that they can subscribe to it for segment updates,
and it's the daemon who is going to trigger those clients through
the write entry point.