Added support for write() and IntrDAEMON
The write() entry point can be used for clients to wake up other
clients subscribed to it.

This allows a certain hierarchy in notifying a particular event; let's see
a little example where we have two different types of clients (A and B).

* Hardware Interrupt -> Clients A (Clients subscribed to it are woken up)
* Clients A (e.g. daemon) do some pre-processing (e.g. write to RAM)
* Clients A issue a write which will wake up Clients B (subscribed to
  IntrDAEMON)
* Clients B are woken up and do their job (e.g. read from RAM).

Note. This code is untested (at the time of writing) because we haven't
still made the modifications in our user space programs (library/daemons)
to use it.