Moved function comments from code to header files
Proceeding this way we can generate the documentation for the
library only, which is what we want to export to the users.

There is a downside, though: now the C code is a bit more difficult
to read, since the comments for each of the functions are not in the
same file.