xmemDrvr.c: IntrHandler, improved comments regarding LISR/LIER
The comments were wrong; the Interrupt Enable bit in LISR is cleared
when LISR is read; however, in order to serve interrupts that are
coming at a fast rate, we re-enable them after Interrupt Enable has
been cleared. Therefore we could serve several almost-at-the-same-time
interrupts without getting out of the Interrupt Handler.