VME: Support for VME modules using the TSI148 VME bus driver
There is no need to keep track of the interrupts and the mappings
inside CDCM; the driver takes care of that, so we don't do anything
there -- hence a lot of (unused and not-working as it was) code can
be removed.

vmebus.h is added temporarily since it's yet to be decided how we
are going to compile driver who use the TSI148 VME bus driver.