Compilation changes.
Compiled driver objects has .ko extention.

All files (not only those, with *Drvr.c suffixes) in the current directory
are considered to be a part of the driver and compiled.
Files with comma (,) prefix are mangled and will not be compiled.

KVER changed to be the output of 'uname -r' command (4.0.0 in Lynx case)