64K-round up/down a mapping *only* upon window creation
vme_create_window() is the right place to round up/down the size
and address of a desired mapping. Hence include both checks in
that function, and remove any prior references to 64K alignment.

Rounding before entering vme_create_window was wrong because the
mapping that the user receives should have the offset and size
that he asked for (another matter is the window where that
mapping lives in).

For example, asking for a mapping whose size and VME address
are both smaller than 64K, would return to the user
the virtual address of a mapping which starts at 0x0 and
ends at 64K, instead of pointing to the right VME offset.

Now, this situation is handled gracefully:

           va       PCI     VME      Size
--------------------------------------------------

Window 0:  Active -  1 user
        f83c0000 81010000 00000000 00010000 - (0x29)A16 USER / D32  NOPF

Mappings:
     0: f83c0400 81010400 00000400 00000400 - (0x29)A16 USER / D32  NOPF

The user gets f83c0400 back, which is what he wants.

Signed-off-by: Emilio G. Cota <emilio.garcia.cota@cern.ch>