InsLibCloneOneDriver() is not performed by CDCM anymore.
It's up to the user to call it from installation entry point.
If the driver is 'Skel' compliant - then SkelDrvrInstall() is doing it.

Address of the info table pointer is passed into install entry point
by the CDCM.
It's done to be fully Lynx-driver compliant.

Note however, that in case of special Linux installation,
install routine is passing the user-space address of the info table
(but __not__ the address of the info table pointer!) in 'infoaddr'
parameter (see do_install_driver() for more details).