cdcmTime: semaphores re-work
This new implementation simplifies the code a lot, and solves the
bugs that were present in the previous implementation (e.g. different
spinlocks to protect the same structure, wrong signaling of the
semaphore for certain values of the semaphore counter.)

This implementation is almost a verbatim copy of the Linux kernel
semaphores merged in 2.6.26 (author: Matthew Wilcox). Some features
have not been implemented (such as down_killable and down_trylock)
because there's no direct LynxOS equivalent, and some others
have been added to provide Lynx' features (such as what happens
with scount.)

Also, the copyright notices in cdcmTime.[ch] have been updated.

Signed-off-by: Emilio G. Cota <emilio.garcia.cota@cern.ch>