cdcmBoth.h: add Linux-like spinlocks and mutexes
The idea is to take advantage of the most common Linux locking
mechanisms; these are wrapped with 'cdcm_' in front so that
we can use the appropriate locking primitives in Lynx.

The implementation is fairly straightforward: all the interrupt-safe
locking mechanisms are implemented in Lynx as disable/restore,
and mutexes are (in Lynx) just semaphores. In Linux they're just
the native calls.

Signed-off-by: Emilio G. Cota <emilio.garcia.cota@cern.ch>