cdcm.h: use struct cdcm_file instead of struct file
So far we've been substituting 'file' by 'cdcm_file', so that
the drivers could keep being Lynx-like, and Linux would see
'struct cdcm_file' instead of 'struct file'. This has a problem:
Linux kernel functions embedded in the code make use
of 'struct file' sometimes, and of course the 'struct file' they
expect is the Linux one.
Therefore the old procedure was broken very easily. The solution
we've found is to use 'struct cdcm_file' in all the driver's entry
points, which will be substituted to the right thing in both
platforms. This comes at the cost of changing the driver's code
a bit, but the gain is that the compilation is much more robust.

Signed-off-by: Emilio G. Cota <emilio.garcia.cota@cern.ch>