semaphores: fix initialisation lockdep-related bug
A lockdep key cannot be a stack variable from a module. We
were doing that and is plain wrong (since it doesn't
guarantee uniqueness). In fact lockdep was triggering it.

Therefore remove the 'broken home-made' lockdep initialisation
for our semaphores; this means we have no lockdep for our
semaphores (although the spinlocks protecting them are
tracked). If we want to have that, we should just wrap
the native semaphores, but this solution would only work
for kernels >= 2.6.26. So leave it like this for the time being.

Signed-off-by: Emilio G. Cota <emilio.garcia.cota@cern.ch>