vme_window: define struct vme_taskinfo to store task attributes
Just storing a pointer to struct task_struct in struct mapping is
bogus, since usually, soon after a mapping is done the task doesn't
exist anymore, but we still keep the pointer hoping that it points
to the right place (for instance when printing the /proc/vme/windows
entry, which in many times causes buffer overflow since we're just
printing garbage).

To fix it, define 'struct vme_taskinfo', which keeps a local copy
of the pid and task name, so that it can be safely used throughout
the code.

Signed-off-by: Emilio G. Cota <emilio.garcia.cota@cern.ch>