tsi148.c: fix outbound ending address miscalculation
The outbound ending address calculation is wrong: the addresses inside
the requested window get translated into the right VME addresses, but
also the addresses in the 64K chunk beyond the end of the requested
window get translated.

This leads to overlap among the target addresses, which is resolved
by the chip by accessing the VME bus as if it were configured by
the overlapping window with the smallest window number. In other words,
installing two different drivers that make use of different address
spaces won't work for one of them, since their addresses will overlap,
at least partially.

This patch fixes it by making sure that the translation range is
as requested; note that for instance, the mapping of 64K (which
is the minimum possible window size) will have its starting
and ending address equal.

Example: TSI148 lcsr outbound translations

Before the patch:
No. otat         otsau:otsal        oteau:oteal         otofu:otofl
0:  80040045  00000000:81010000  00000000:81090000:  FFFFFFFF:7F570000
1:  80040040  00000000:81090000  00000000:810A0000:  FFFFFFFF:7EF70000

Note the overlap between windows 0 and 1 since the ending address of 0
is equal to the starting address of 1.

patched:
No. otat         otsau:otsal        oteau:oteal         otofu:otofl
0:  80040045  00000000:81010000  00000000:81080000:  FFFFFFFF:7F570000
1:  80040040  00000000:81090000  00000000:81090000:  FFFFFFFF:7EF70000

Note that the starting address of window 1 is different to the
ending address of 0--hence no clash. Also note the starting and ending
address of 1 are the same; i.e. the size of 1 is 64K, as requested.

Signed-off-by: Emilio G. Cota <emilio.garcia.cota@cern.ch>