skel-isation
Substitute the original driver by a new, skel-based one.
This saves code and solves a few problems:
 - incorrect/non-existent locking while connecting/disconnecting:
   skel takes care of that now.
 - ensuring atomicity while executing MTT commands: a spinlock
   has been added for this.
 - code duplication among read(), write() and IntrHandler()

The changes above done to the driver require user-space to
accommodate them:
 - IOCTL names/numbers have changed
 - many structs have changed from MttXXX to SkelXXX
Therefore the library and test programs are also modified
by this commit.

Since the kernel-userspace interface has changed, the new
driver is not compatible anymore with previous user-space
code; that's the reason why it's been renamed from 'mtt'
to 'mttn' ('n' referring to 'new', or 'next'.)

Signed-off-by: Emilio G. Cota <emilio.garcia.cota@cern.ch>