build: solve MttDrvr* include madness
Many user-space code, which should only use the library,
used to include mttdrvr.h; now with the additions of the
skel definitions, this becomes less self-contained and
therefore uglier (including a driver header file from code
that only uses the library was the first mistake).
Attempt to improve the situation a bit by defining an mtt
header file, which is meant to be included by everyone,
and doesn't have any skel crap in it.

Note: a user-space application that wants to make direct
calls to the driver, only needs to include drvrmtt.h,
as it used to.

Signed-off-by: Emilio G. Cota <cota@braap.org>