driver: take usleep() out of the atomic region in reset()
This actually opens a possible window for accessing the module
just after it's been reset, but having interrupts disabled
while waiting for anything is a much bigger sin.

We could try to be clever by defining a spinlock around every
I/O operation as well as the iolock spinlock, but this would
probably be an overkill, so probably leaving it like that is
good enough.

Signed-off-by: Emilio G. Cota <cota@braap.org>