Add ModuleName in InsLibModlDesc, modularize the code, cleanup bugs
As each module has the name in the DB, and this name is needed
(by the driverGen for ex) it's a good idea to have a <name> tag
in .xml driver description.

Before module name was passed into <driver> tag 'name' field -- which is
wrong (Emilio spotted this one).
Same driver can be used by several modes (why not).
For ex. XMEM driver is used by VMIPMC-5565 module.
If we put VMIPMC-5565 name in the <driver> 'name' field -- installation
program will try to search for the wrong driver.

Now drivers and modules has their own real names.

Signed-off-by: Yury Georgievskiy <ygeorgie@cern.ch>