tsi148.c: use scatter-gather DMA irrespective of the transfer size
DMA transfers to/from buffers that do not cross a page boundary
are currently done in single mode. This is likely to fail when
(M)BLT mode is enabled, since transferring the buffer as is
would probably cross a certain boundary in the VME bus.

Fix this by deferring always to the linked list code path,
which handles this gracefully, therefore removing the code
that performs single mode transfers. This means that in some
cases (small transfers) the overhead will be slightly greater,
since DMA descriptors need to be allocated from the DMA pool
and filled in. This penalty is outbalanced by the fact that we
only have one DMA code path to maintain instead of two, i.e.
less bugs and quicker testing.

And anyway, for the small transfer sizes in which this impact
would be noticeable, the caller shouldn't be using DMA at all.

Signed-off-by: Emilio G. Cota <cota@braap.org>