tsi148.c: mark the invalid use of tsi148_dma_link_add() with BUG_ON
**virt is in this function a pointer passed by reference
which points to the current DMA descriptor. This function
then updates the pointer to point to the next descriptor,
or sets it to NULL if this was the very last link in the chain.

It is therefore a sign of breakage to pass this pointer as
NULL, since that would mean that the caller is requesting
a buffer overflow.

Signed-off-by: Emilio G. Cota <cota@braap.org>