DMA: Abort DMA transfers when receiving a signal
User-space processes may be signalled while waiting on
the completion of a DMA transfer. This is not properly
handled: the VME driver does not check the return value from
wait_interruptible(), consequently unmapping memory which the
bridge might still be accessing asynchronously.

This bug is easily triggered by interrupting (Ctrl-C)
hsm-stress.sh:

[cfv-864-cdv7] ./hsm-stress.sh -d10 -n4
executing task 1
executing task 2
executing task 3
executing task 4
^C
[ then stats are printed..]

And the following messages show up in the kernel log:

[  120.117792] VME Bridge: PCI error at address: 00000000 07046000 - attributes: c000400c
[  120.117795] PCI-X attributes: 60000004 - PCI-X split completion: 10000100
[  120.299541] VME Bridge: PCI error at address: 00000000 03046000 - attributes: c000400c
[  120.299543] PCI-X attributes: 60000004 - PCI-X split completion: 10000100

This bug can have pretty nasty consequences: it sometimes hard
freezes the box.

The bug is in the wait_interruptible() call, which puts the
process to sleep until the DMA transfer finishes: the return
code is not checked, thus not handling signals properly.

Fix this by checking the return value from wait_interruptible()
immediately: if a signal interrupts the process, the ongoing
DMA operation is aborted; after aborting, the mappings created
for this DMA operation are cleaned up safely as usual.

Signed-off-by: Emilio G. Cota <cota@braap.org>