test: add a DMA test program using the HSM 8170 from CES
This patch adds support for testing DMA on the HSM 8170
VME memory from CES. The test is only concerned with DMA
performance: the board is accessed constantly via DMA without
any mappings done.

On top of this test program, a little script is also added in
this patch to launch several instances of the test program
concurrently. The goal here is to debug any concurrency
issues and have a quick way of stressing the DMA engines in
the VME bridge.

Note that this test is not concerned with data coherency
at all.  In fact, the board seems to have a problem there
in BLT mode.  Check the comment added at the top of the test
program source for more information on this.

Signed-off-by: Emilio G. Cota <cota@braap.org>