vme-sysfs: export a parameter to show bus errors on the kernel log
By default, the VME bus error flag isn't enabled. That means
that VME bus errors do not show up on the kernel log.
This is not wrong 'per se': imagine a crazy user-space task
generating lots of VME bus errors. Since these messages would
be passed to the serial console as well, printing them all
would slow down the box dramatically.

However, when a crate is properly configured, from time to time
(normally due to a hardware error) we might get a bus error, and
we certainly want to leave a trace of these in the kernel log.

For these cases this patch provides a sysfs knob to enable the
printing of VME bus errors on the kernel log.

Example: user-space task accessing bogus VME addresses..

[cfv-864-cdv7]  /  > echo 0 > /sys/module/vmebus/parameters/vme_report_bus_error
[cfv-864-cdv7]  /  > echo 1 > /sys/module/vmebus/parameters/vme_report_bus_error

[  237.855712] VME Bridge: VME bus error at address: 00000000 00010010 - AM: 0x9 - tsi148 attributes: 800ec900
[  238.602238] VME Bridge: VME bus error at address: 00000000 00010014 - AM: 0x9 - tsi148 attributes: 800ec900
[  239.603766] VME Bridge: VME bus error at address: 00000000 00010018 - AM: 0x9 - tsi148 attributes: 800ec900
[...]

Signed-off-by: Emilio G. Cota <cota@braap.org>