vme: Check for VME bus errors given by their Address and AM
The existing bus error checking interface is insufficient,
since it is likely to cause clashes between two tasks which
concurrently check (and clear) bus errors.

Fix this by adding a new symbol [ Note. we do not want to change
the interface of the old one, since there are drivers out
there using it ] which takes the address and adress modifier of
a bus error candidate as parameters. If the last bus error seen
on the bus matches this candidate, it is cleared and reported.
Otherwise it is left there until there's a match or a new bus error
arrives and overwrites it.

Signed-off-by: Emilio G. Cota <cota@braap.org>