vme_bridge.c: Do not warn when sitting on non-VME64x backplanes
When the board is plugged on a non-VME64x backplane, the
Geographical Address (GA) signal will be such that the slot
number that the board reads is 0.

Handle this properly instead of giving the cryptic (and bogus)
"Weird VME slot" message when this happens. This message is
kept as a warning for the cases where the VME slot is not 0 nor
[1,21], which indeed points to a faulty backplane and should
be notified.

Signed-off-by: Emilio G. Cota <cota@braap.org>