lib: remove redundant parameter of vme_bus_error_check_clear()
The signature of vme_bus_error_check_clear is unnecessarily
long and ugly, since the address modifier of the bus error to
be checked can be retrieved from the vme_mapping struct which
is already passed as a pointer.

This patch removes this redundant parameter.

Note that there are no users (yet) of this function, so
we're changing the API now before the design mistake becomes
irremediable.

Signed-off-by: Emilio G. Cota <cota@braap.org>
Reported-by: Stephen Page <Stephen.Page@cern.ch>