vme_window.c: keep track of a mappings' ownership via struct file
When a process requests more than one mapping through vme_map(),
all the subsequent calls to vme_unmap() fail except the first
one. This is a bug in the vme driver that can be triggered
with test/doublemap.c.

The driver keeps track of the owner of the mapping by its
pid. When a process closes the vme window file descriptor,
which is done by vme_unmap(), all the mappings of that pid
are released.

Fix this by associating each mapping coming from user-space with
the resulting struct file in the kernel. This way for each call
to vme_map() there is a corresponding struct file (resulting
from open()) in the kernel.  If a process is killed, all the
file descriptors it opened will be closed and all the mappings
will be undone, avoiding any leakage of mappings.

While at it, do not show in /proc the pid and task name of
mappings which were requested by kernel code. For those, store
'kernel' as the client, as opposed to 'insmod' and its pid,
since that process doesn't exist anymore, rendering that
information useless.

Before the patch:

$ cat /proc/vme/windows | grep -A6 'Window 7'
Window 7:  Active -  1 user
        c0700000 80010000 00000000 01000000 - (0x39)A24 USER DATA / D32  NOPF

  Mappings:
     0: c16ff000 8100f000 00fff000 00001000 - (0x39)A24 USER DATA / D32  NOPF
        client: 1189 insmod

After:

$ cat /proc/vme/windows | grep -A5 'Window 7'
Window 7:  Active -  1 user
        c0700000 80010000 00000000 01000000 - (0x39)A24 USER DATA / D32  NOPF

  Mappings:
     0: c16ff000 8100f000 00fff000 00001000 - (0x39)A24 USER DATA / D32  NOPF
        client: kernel

Signed-off-by: Emilio G. Cota <cota@braap.org>
Reported-by: Maxim Andersen <Maxim.Andersen@cern.ch>
Reported-by: Stephen Page <Stephen.Page@cern.ch>
Tested-by: Maxim Andersen <Maxim.Andersen@cern.ch>
Tested-by: Stephen Page <Stephen.Page@cern.ch>