lib: allow the mapping of non-page-aligned VME regions
vme_map() requires page-aligned VME regions to work.  Some cards
cannot then be mapped to user-space because they cannot always
be set at page-aligned offsets.

Remove this requirement by mapping in vme_map() a page-aligned
region which contains the requested (non-aligned) region. The
same rounding is done in vme_unmap() so that this is transparent
to the callers.

Signed-off-by: Emilio G. Cota <cota@braap.org>
Reported-by: Stephen Page <Stephen.Page@cern.ch>