utils/extest: use GNU readline instead of lite_shell
Using our own lite_shell for processing stdint in extest was
a Wrong Idea(tm) for two reasons:

1. It doesn't work--in Lynx, it segfaults like hell.

2. Go to 1.

One way of fixing it might be to go to the code and try to
work out what's going on. But this is not a good idea, either:
the code is tricky and painful, as it usually is with curses.

The Right Way(tm) of fixing was long ago pointed out by the
Finnish demigods:

"Intelligence is the ability to avoid doing work, yet getting
the work done."

- Linus Torvalds

In our case this means we should leave the pain of dealing with
Curses & Family (now that I think about it, it makes sense they
called it _Curses_) for people who are better at it than us.

This gets rid of a lot of code while giving us more features
(beat _that_[1], lite_shell!).

And most importantly, It Just Works. [*]

[1] http://tiswww.case.edu/php/chet/readline/rluserman.html

[*] I've tested vd80Tst and vd80libTst on ppc4 and L865.

Final note: special mention to Nicolas (nmn), he's taken the
pain of compiling and installing the latest version (6.1)
of GNU readline on lynx just for our pleasure.

Signed-off-by: Emilio G. Cota <cota@braap.org>
Reported-by: Matthieu Cattin <Matthieu.Cattin@cern.ch>