skel: convert ModuleConnection's rwlock into spinlock
The usage of rwlocks instead of spinlocks here isn't justified.
read_lock() is only used a couple of times, on paths that are
never hit by many readers. So really there's no point in using
rwlocks--in fact it is probably slower than using spinlocks. And
anyway, if what we wanted was a mechanism for dealing with
many concurrent readers, we should have used RCU from day one.

This converts ModuleConnection's rwlock into a spinlock, using
the irqsave/restore version everywhere--it's safer and simpler
(with a negligible performance hit) since critical isr/non-isr
code paths are shared.

Signed-off-by: Emilio G. Cota <cota@braap.org>