vme: refuse to map regions whose base address is out of range
Currently, it is possible to map a region whose base
address sticks out of its eventual address space (e.g. a
24-bit-long base address for an A16 mapping). This is a sign
of misconfiguration and may have unpredictable results.

Fix this by detecting the invalid address, refusing to map
the region.  A warning message is printed to the kernel so that
the user will be able to identify the problem and (hopefully)
fix it.

Signed-off-by: Emilio G. Cota <cota@braap.org> Reported-by:
Juan David Gonzalez Cobas <dcobas@cern.ch>