driver/Makefile: build as a kosher Linux kernel module
This splits the previous Makefile into a Kbuild file and
a Makefile.

The point is to use all the goodies we can get from the kernel,
such as `make help'. Thanks to kbuild the Makefile turns out
to be a mere wrapper around `make -C /path/to/kernel_sources'.

With this patch it's really easy to compile the driver for
non-CERN-related targets; as long as the path to the kernel
sources in common.mk is valid, `make' does it all.

Signed-off-by: Emilio G. Cota <cota@braap.org>