skel: add get_vme_addr() to retrieve the VME address of a module
This is useful for drivers which want to retrieve the VME address
of a region. This is needed usually when DMAing to/from a module,
where the address of the module on the VME bus has to be provided.

Drivers that use DMA, such as vd80 or the upcoming sis33, should
use this helper instead of open-coding it.

Signed-off-by: Emilio G. Cota <cota@braap.org>