DMA: add vme_do_dma_kernel()
This function is used to DMA to/from kernel buffers.

The code behind it is almost entirely shared with vme_do_dma,
which does DMA to/from user-space buffers. The only difference
is in the way the pages of the buffer are retrieved.

Note that we use virt_to_page(), which only works for lowmem.

Signed-off-by: Emilio G. Cota <cota@braap.org>