coht/vmebridge: use coht's makefiles
This is the first commit after setting up vmebridge/ to track
the vmebridge project as a subtree. Before this commit,
there are two commits that I'm omitting:

- one that removes the coht/vmebridge directory
- one that merges the vmebridge directory as a subtree.
  For more info on the subtree strategy, see:
http://www.kernel.org/pub/software/scm/git/docs/howto/using-merge-subtree.html

This way we can track the project, pulling with

$ git pull -s subtree VME_REMOTE master,

where VME_REMOTE is the remote that points to the vmebridge project.

The cool thing is that we can keep separate makefiles here, which
is what this commit does.

Of course, if we try to pull from the original project changes
that touch the original makefiles, we'll get conflicts. This is
unavoidable, but they shouldn't be difficult to resolve.

I think this is the best we can do without going mental.
Thoughts?

Signed-off-by: Emilio G. Cota <cota@braap.org>