skel: Make skel use only a single node in /dev
In this mod client contexts are allocated dynamically, there is still a restriction on the maximum
number of clients due to using bits in a uint32_t to represent a client mask. We also still need
to put interrupts on client queues, so an array of client context pointers is kept in the work area.
Here the extended cdcm_file structure keep a pointer to the client context in the ->buffer pointer.

Signed-off-by: Julian Lewis <julian.lewis@cern.ch>