[utils] [cdcm] Fix sysbrk dependency in core components
sysbrk belongs to either LynxOS code or Linux emulation of LynxOS
system calls. The right place for it is, therefore
    a) as part of cdcm
    b) as part of code that must run on both platforms
that is, libinstkernel.

So cdcm implements sysbrk now as part of its LynxAPI part
(as promised in cdcmLynxAPI.h), libinstkernel has its own
static/private implementation of it, and the former implementation
in drvr_utils.c is removed.

This mod does not break any compilation, as far as observed.

En passant, a bug is fixed in former implementation of sysbrk/sysfree.
The boundary for going to vmalloc instead of kmalloc is set to 128KB;
it was (wrongly) set to 128MB.

Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>