utils: add name property to CAR section of drivers.xml
To accurately locate a mezzanine in a crate, we need to know
what card it is plugged in, and which slot in that carrier board
it is lying to. This last parameter is the board_position property
at a CAR section of drivers.xml.

However, specification of the carrier board  is not completely
achieved by giving a board_number property at that CAR section,
because this board_number is an index within a family of carrier
boards (actually, a driver) that could vary. For example,
a MODULBUS carrier can be a VMODIO, MOD-PCI or MOD-PCI4 card.
To accurately position a VMOD12A2, we need to specify

	name = "MOD-PCI"
	board_number = "2"
	board_position = "3"

and the name is needed to disambiguate whether board number
2 is the second mod-pci or the second mod-pci4 in the system.

This patch takes care of adding this (to date) absent field

Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>
Signed-off-by: Juan David Gonzalez Cobas <david.cobas@gmail.com>