utils: accept MODULBUS and IP as valid carrier bus names
libinst expects that carrier boards are described in the XML
file by a bus_type="CAR" attribute. This is in disagreement with
what the CCDB provides, which is the true bus name of the carrier
family (e.g., MODULBUS, IP for Industry Pack, etc.).

With this modification, both types of carrier buses are
accepted just as a "CAR" tag is.

Signed-off-by: Juan David Gonzalez Cobas <david.cobas@gmail.com>