mtt: fix task loading bug on little endian machines
The driver was assuming that a task to be loaded was just a blob
of u32's, and was therefore swapping each u32 on LE machines.

A task is an array of instructions. An instruction is described as:

(from include/asm.h:)
typedef struct {
   unsigned char  Dest;   /* Destination reg in most significant byte */
   unsigned char  Src2;   /* Second source reg low byte of high short */
   unsigned char  Number; /* Opcode number in high byte of low short  */
   unsigned char  Crc;    /* Check sum in the least significant byte  */
            long  Src1;
 } Instruction;

For some obscure reason, the driver sees an instruction as

(include/mtthard.h)
typedef struct {
        unsigned long Fields; /* Encoded bytes: Dest, Srce2, Numb, Crc */
        long Srce1;  /* Type Source register/value */
} MttDrvrInstruction;

, and was swapping both u32's when accessing the module.

This patch handles this more carefully, not swapping the first field,
which is the right thing.

Signed-off-by: Emilio G. Cota <cota@braap.org>
Reported-by: Jean-Claude Bau <Jean-Claude.Bau@cern.ch>