doxy: add doxygen automation scripts
When generating documentation with doxygen, in most cases
we just change a few parameters: project name, input files
and output directory.  The rest of the doxygen config file
remains the same, and therefore should be shared among all
our sub-projects.

AFAIK a doxygen config file cannot include other doxygen
config files, so I've come up with a script that changes the
parameters above on a pre-defined config file, piping the
result to doxygen.

This generates doxygen output in man, html, and latex formats.

Signed-off-by: Emilio G. Cota <cota@braap.org>