extest: increase the maximum number of atoms in a line
'640K of memory should be enough for anybody.'
 - Bill Gates, at some point in 1981

Well, the quotation above doesn't come from Bill Gates--he
denies he ever said such a stupid thing [1]. However, it
illustrates the problem of setting an upper limit for anything;
there always comes the time when that limit becomes obsolete.

The time to revise the maximum amount of arguments (atoms)
in a line has come for extest, too; Matthieu is debugging one
of his boards with the test program, and needs at least ~150
arguments to be processed in a single line.

This patch sets it to 256, "which should be enough for anybody,
until it isn't enough anymore."

[1] http://www.wired.com/politics/law/news/1997/01/1484

Signed-off-by: Emilio G. Cota <cota@braap.org>
Reported-by: Matthieu Cattin <Matthieu.Cattin@cern.ch>