cdcm + ctr/cdcmSem: swait(): return either OK or SYSERR
From LynxOS' manual:
    swait returns SYSERR if flag is SEM_SIGABORT and a signal
    interrupts the wait, otherwise it returns OK.

From the description above, we currently comply only with
the second part; that is we do return OK when the semaphore
is acquired.

However, when the semaphore isn't acquired, we don't return
SYSERR; after a few functions get called, we eventually return
either ETIME or EINTR, which come from __cdcm_down_common():

@@ __cdcm_down_common() [...]
 timed_out:
	list_del(&waiter.list);
	return -ETIME;

 interrupted:
	list_del(&waiter.list);
	return -EINTR;

This patch fixes this by returning OK when the semaphore has
been acquired and SYSERR otherwise.

Note that this bug in the implementation hasn't ever been
noticed because we usually wait on semaphores as follows:

	if (swait(semaphore, SEM_SIGABORT)) {
		could_not_acquire_semaphore();
		return -1;
	}

Signed-off-by: Emilio G. Cota <cota@braap.org>
Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>