skel: fix wrong check of empty list in __fill_client_queues
The value of *client_list won't ever be NULL, because it is
the address of a field within a struct that exists in memory.

What the program must check at that point, which is what it
was doing before commit a399289, is that no one is connected
to interrupt 'i'.

Signed-off-by: Emilio G. Cota <cota@braap.org>