skel: move GET_CLIENT_LIST handler to its own helper
This code was too ugly to live, with heavy nesting and a
horrible goto as its main sins.
This makes it a little bit more palatable by using helper
functions whenever possible.

Note that the module number assigned is now mcon->ModuleNumber
instead of the index within the ModuleContexts[] array, which
was wrong because skel accepts non-contiguous module numbers.

Signed-off-by: Emilio G. Cota <cota@braap.org>