skel: fix scheduling while atomic bug in get_add_client()
get_add_client() allocates memory and is called from atomic
sections (i.e. spinlocks held), which is a no-no.

This patch eliminates this function, implementing its
functionality directly in the original caller.

Now the allocation is done in process context. Then in the
atomic region we may or may not use the allocated link.
If we don't use it, we free it.

I know this is a bit complex; it would be much easier to do
kmalloc(..,GFP_ATOMIC), but Lynx doesn't have that ;-(

Signed-off-by: Emilio G. Cota <cota@braap.org>
Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>