skel: eliminate remove_client() and fix the interrupt-disabling bug
remove_client(), which has a single caller, does very simple
stuff.  It's so simple that it doesn't qualify to be in a
separate function--in fact one could argue that it's worse
because the caller may wonder if it does any locking or not.
This patch removes the function.

This also fixes a bug, reported by Julian, which causes a
module to disable a particular interrupt every time a client
disconnects from it, even if there are other clients connected
to the same interrupt.

Reported-by: Julian Lewis <julian.lewis@cern.ch>
Signed-off-by: Emilio G. Cota <cota@braap.org>
Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>