cdcm: get rid of the "CDCM memory allocator" (fix {r,w}bounds bug)
Yesterday Julian reported the following bug: when two processes
issue concurrently a system to a CDCM driver, we see messages like
the following in the kernel log:

[cvorg_WARN] SkelDrvrIoctl: ILLEGAL NON NULL ARG, RCNT=0/8, WCNT=0/8, cmd=-1073199344 (Get SRAM) pid=2579

That is printed out when either rbounds or wbounds fail.  Looking
into it I figured out the implementation of these two was racy,
and that there was a lot of stuff underneath that we could remove.

CDCM keeps a list of allocated memory blocks, each of them having a
descriptor to store the size and flags of each block.  Most blocks
are pre-allocated when installing the driver.

The block descriptors are stored in a linked list. Unfortunately
this list doesn't have any locking, which is the cause of the bug.

Instead of adding locking and complicating things even more, this
patch does the following:

- Removes the 'CDCM memory allocator'--there's no reason to use
  pre-allocated memory blocks.

- Allocates memory dynamically for each read(2), write(2) and
  ioctl(2) and then deallocates it.

- The flags and size of each of the buffers are prepended to them,
  making the implementation of the whole thing almost trivial.

After applying the patch the bug seems to be gone.

Reported-by: Julian Lewis <julian.lewis@cern.ch>
Signed-off-by: Emilio G. Cota <cota@braap.org>
Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>