cdcm: r/w driver-gen functions relies on CDCM to handle copy_to/from_user
driver-gen UEP were called before CDCM allocates memory and checks its boudnaries.
Now it relies on cdcm to do all checks needed, and is called after memory was
allocated by the cdcm.

ioctl copy_to/from_user is still managed by the driver-gen as specific
information is passed to it in the call and processed by the driver-gen.

Signed-off-by: Yury Georgievskiy <ygeorgie@cern.ch>
Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>