mtt/lib: GetRunningTasks(): fill in NUL as the task name for non-running tasks
Callers of this function provide an array where the names of
the currently running tasks are filled in.

A stopped task should be marked as such with NUL, as in
MttLibGetLoadedTasks(); that is how the caller discerns if
the task meets the condition (running, stopped, etc.) or not.

This isn't done in GetRunningTasks(), which is the bug that
this patch fixes.

Reported-by: Jean-Claude Bau <Jean-Claude.Bau@cern.ch>
Tested-by: Jean-Claude Bau <Jean-Claude.Bau@cern.ch>
Signed-off-by: Emilio G. Cota <cota@braap.org>