makefiles: Add default target handler
Default target for all framework subdirs is 'all'
User is calling this target to build src code and can provide
CPU he wants to compile for.
It is checked for validity, and local '_build' rule will be called
for all needed CPU's.

If provided CPU is 'default' -- compilation will be called for
default CPUs, which are L865 and ppc4

Signed-off-by: Yury Georgievskiy <ygeorgie@cern.ch>