vd80: Fix analog trigger mistakes
This commit fixes a couple of errors in vd80 analog trigger
configuration.

First, we want to trigger on edge, not on level, so that triggers are
generated when the signal enters the configured region. This requires
setting the ATRIGCHANGECHn bit to 1 in the channel Analog Tirgger Level
register (0xC0..0xFC).

Second, the bit ATRIGCHANGECHn is 4, not 3. This mistake often happens
when people do not count from 0, as most normal humans do since K&R.

Signed-off-by: Juan David Gonzalez Cobas <david.cobas@gmail.com>