makefiles: Add ROOTDIR && use TARGET_CPU instead of CPU in Makefile.specific
Driver framework rootdir (where cdcm, utils etc. is located) should be defined
in driver-specific makefile.
By default it is set to '..' (in examples/Makefile.specific)

TARGET_CPU is used instead of CPU to distinguish between CPU that was provided
in the command line as a parameter and one that was set within Makefile.specific

It in turn is used by rules.mk to compile for all needed CPUs

Signed-off-by: Yury Georgievskiy <ygeorgie@cern.ch>