vd80: program around vd80 VDAT-1 bug
When a STOP command is issued, the vd80 takes some clock-dependent time
to settle in IDLE state. This (VDAT-1) issue results also in a spurious
acquisition interrupt to appear (codenamed VDAT-6 in vd80's Jira
record).

This is a workaround that forces us to wait for an IDLE state to appear
before getting further.

Signed-off-by: Juan David Gonzalez Cobas <david.cobas@gmail.com>